% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsn.r
\name{permutation_matrix}
\alias{permutation_matrix}
\title{generate the permutation matrix from the reduced word}
\usage{
permutation_matrix(word, k)
}
\arguments{
\item{word}{reduced word}

\item{k}{position}
}
\value{
signed permutation matrix
}
\description{
return the signed permutation matrix corresponding to the
word
}
