% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.r
\name{shifted_hook_walk_xtrace}
\alias{shifted_hook_walk_xtrace}
\title{Sample max-cell x trace from shifted SYT}
\usage{
shifted_hook_walk_xtrace(n)
}
\arguments{
\item{n}{size parameter}
}
\value{
integer vector
}
\description{
Sample max-cell x trace from shifted SYT
}
