% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsn.r
\name{plot_trajectory}
\alias{plot_trajectory}
\title{Plot particle trajectories}
\usage{
plot_trajectory(traj)
}
\arguments{
\item{traj}{trajectory object}
}
\description{
Plot particle trajectories
}
\examples{
w <- shifted_hook_walk_xtrace(30)
tr <- trajectory(w, particles = c(1, 5 * (1:6)))
plot_trajectory(tr)
}
